pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "minigpt"
version = "0.1.0"
description = "A modular, research-oriented decoder-only Transformer framework for pretraining, fine-tuning, aligning, evaluating, quantizing, and deploying small GPT models."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "MiniGPT Authors"},
]

dependencies = [
    "torch>=2.4",
    "transformers>=4.40",
    "datasets>=2.18",
    "tokenizers>=0.15",
    "wandb>=0.16",
    "trl>=0.8",
]

[project.optional-dependencies]
eval = [
    "lm-eval>=0.4",
    "google-cloud-aiplatform>=1.50",
]
quant = [
    "autoawq>=0.2",
    "auto-gptq>=0.7",
    "llama-cpp-python>=0.2",
]
serve = [
    "vllm>=0.4",
    "sglang>=0.1",
    "gradio>=4.0",
]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "ruff>=0.3",
]
all = ["minigpt[eval,quant,serve,dev]"]

[tool.setuptools.packages.find]
include = ["miniGPT*", "configs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 100
